fix(backend): fix CI failures introduced by the penalty-snapshot commit
Two self-inflicted bugs surfaced by CI on the previous commit:

1. Migration 00025 declared penalty_assignments.amount as the legacy
   NUMERIC(10,2) instead of BIGINT (integer cents), the type every other
   amount column uses since migration 00008. NUMERIC(10,2) can't hold
   maxAmountCents (100_000_000, i.e. the app-allowed $1,000,000.00
   penalty) without a numeric field overflow -- so a penalty at the top
   of the allowed range could never actually be assigned to anyone, and
   the migration's own backfill UPDATE could fail outright against any
   database already holding such a penalty. Fixed to BIGINT; not yet
   applied anywhere outside this open PR, so edited in place rather than
   via a follow-up migration.

2. CreateAssignment's INSERT sourced amount/label via a scalar subquery
   on penalties keyed by penalty_id. Against a deleted penalty, that
   subquery evaluates to NULL rather than failing -- which tripped the
   NOT NULL constraint (an immediate, per-row check) before the
   penalty_id FK violation (a constraint trigger, checked after the row
   is built) ever got a chance to fire, turning the intended
   ErrPenaltyNotInTeam into an unmapped "violates not-null constraint"
   error. Rewrote as a two-step read-then-insert inside a transaction:
   read the penalty's current amount/label first (returning
   ErrPenaltyNotInTeam directly if it's already gone), then insert with
   real non-null values so a race between the read and the insert still
   surfaces as the FK violation.

Also fixes a test bug (not a real security bug -- the escalation guard
itself worked correctly, confirmed by the audit log in the CI failure):
TestRolesRepository_UpdateRole_EscalationBeyondCallersOwnPermissions_Blocked
asserted the untouched role's Finances permission equals "none", but the
role was seeded with only Settings explicitly set, so Finances held Go's
zero-value "" rather than "none". Seeded explicit "none" values instead
of relying on the zero-value.

Verified against a real local PostgreSQL 16 instance (Docker/testcontainers
unavailable in this sandbox) rather than relying on go vet/build alone,
since both bugs only manifest against a live database.

+++ backend/internal/db/migrations/00025_penalty_assignment_amount_snapshot.sql
@@ -12,7 +12,12 @@
 -- assignment row at creation time makes each assignment an immutable record
 -- of what was actually assigned, matching how contributions already store
 -- their own per-row amount instead of joining a shared definition.
-ALTER TABLE penalty_assignments ADD COLUMN amount NUMERIC(10,2);
+-- BIGINT (integer cents), matching every other amount column since 00008
+-- converted them off NUMERIC(10,2) precisely to avoid float/decimal
+-- boundary issues -- not NUMERIC(10,2), which can't even hold
+-- maxAmountCents (100_000_000, i.e. the app-allowed $1,000,000.00 penalty)
+-- without a numeric field overflow, since NUMERIC(10,2) caps out just under 10^8.
+ALTER TABLE penalty_assignments ADD COLUMN amount BIGINT;
 ALTER TABLE penalty_assignments ADD COLUMN label TEXT;
 
 -- Backfill: existing assignments have no historical amount to recover, so
